third_party/blink/web_tests/external/wpt/css/css-sizing/stretch/block-height-002.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] main > div > div 2
  assert_equals: \n<div style="height: stretch; margin: 10px;" data-expected-height="200"></div>\nheight expected 200 but got 180
Harness: the test ran to completion.
